Moved computed plugin into the core library
