+++ src/EcsRx.Benchmarks/EcsRx.Benchmarks.csproj
@@ -9,7 +9,6 @@
       <ProjectReference Include="..\EcsRx.Examples\EcsRx.Examples.csproj" />
       <ProjectReference Include="..\EcsRx.Infrastructure\EcsRx.Infrastructure.csproj" />
       <ProjectReference Include="..\EcsRx.Plugins.Batching\EcsRx.Plugins.Batching.csproj" />
-      <ProjectReference Include="..\EcsRx.Plugins.Computeds\EcsRx.Plugins.Computeds.csproj" />
       <ProjectReference Include="..\EcsRx.Plugins.Persistence\EcsRx.Plugins.Persistence.csproj" />
       <ProjectReference Include="..\EcsRx.Plugins.ReactiveSystems\EcsRx.Plugins.ReactiveSystems.csproj" />
       <ProjectReference Include="..\EcsRx.Plugins.Views\EcsRx.Plugins.Views.csproj" />
@@ -18,7 +17,7 @@
     <ItemGroup>
       <PackageReference Include="BenchmarkDotNet" Version="0.13.12" />
       <PackageReference Include="NSubstitute" Version="5.1.0" />
-      <PackageReference Include="SystemsRx.Infrastructure.Ninject" Version="7.1.59" />
+      <PackageReference Include="SystemsRx.Infrastructure.Ninject" Version="8.0.62" />
     </ItemGroup>
 
 </Project>
