+++ src/EcsRx.Examples/EcsRx.Examples.csproj
@@ -13,8 +13,8 @@
     <PackageReference Include="Persistity.Serializers.LazyData.Json" Version="2.0.47" />
     <PackageReference Include="Spectre.Console" Version="0.48.0" />
     <PackageReference Include="System.Reactive" Version="6.0.0" />
-    <PackageReference Include="SystemsRx.Infrastructure.Ninject" Version="7.1.59" />
-    <PackageReference Include="SystemsRx.ReactiveData" Version="7.1.59" />
+    <PackageReference Include="SystemsRx.Infrastructure.Ninject" Version="8.0.62" />
+    <PackageReference Include="SystemsRx.ReactiveData" Version="8.0.62" />
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\EcsRx.Infrastructure\EcsRx.Infrastructure.csproj">
@@ -22,7 +22,6 @@
       <Name>EcsRx.Infrastructure</Name>
     </ProjectReference>
     <ProjectReference Include="..\EcsRx.Plugins.Batching\EcsRx.Plugins.Batching.csproj" />
-    <ProjectReference Include="..\EcsRx.Plugins.Computeds\EcsRx.Plugins.Computeds.csproj" />
     <ProjectReference Include="..\EcsRx.Plugins.Persistence\EcsRx.Plugins.Persistence.csproj" />
     <ProjectReference Include="..\EcsRx.Plugins.Views\EcsRx.Plugins.Views.csproj" />
     <ProjectReference Include="..\EcsRx\EcsRx.csproj" />
