+++ src/EcsRx.Examples/ExampleApps/ComputedGroupExample/ComputedGroups/LowestHealthComputedGroup.cs
@@ -2,10 +2,10 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Reactive.Linq;
+using EcsRx.Computeds.Groups;
 using EcsRx.Entities;
 using EcsRx.Examples.ExampleApps.ComputedGroupExample.Extensions;
 using EcsRx.Groups.Observable;
-using EcsRx.Plugins.Computeds.Groups;
 
 namespace EcsRx.Examples.ExampleApps.ComputedGroupExample.ComputedGroups
 {
