+++ src/EcsRx.Infrastructure/EcsRx.Infrastructure.csproj
@@ -14,6 +14,6 @@
     <ProjectReference Include="..\EcsRx\EcsRx.csproj" />
   </ItemGroup>
   <ItemGroup>
-    <PackageReference Include="SystemsRx.Infrastructure" Version="7.1.59" />
+    <PackageReference Include="SystemsRx.Infrastructure" Version="8.0.62" />
   </ItemGroup>
 </Project>
