--- src/EcsRx.Plugins.Computeds/ComputedsPlugin.cs
@@ -1,21 +1,0 @@
-using System;
-using System.Collections.Generic;
-using SystemsRx.Infrastructure.Dependencies;
-using SystemsRx.Infrastructure.Plugins;
-using SystemsRx.Systems;
-
-namespace EcsRx.Plugins.Computeds
-{
-    public class ComputedsPlugin : ISystemsRxPlugin
-    {
-        public string Name => "EcsRx Computeds";
-        public Version Version { get; } = new Version("1.0.0");
-        
-        public void SetupDependencies(IDependencyRegistry registry)
-        {
-            // Nothing needs registering
-        }
-        
-        public IEnumerable<ISystem> GetSystemsForRegistration(IDependencyResolver resolver) => Array.Empty<ISystem>();
-    }
-}
