--- src/EcsRx.Plugins.Computeds/EcsRx.Plugins.Computeds.csproj
@@ -1,24 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-    <PropertyGroup>
-        <Version>0.0.0</Version>
-        <TargetFramework>netstandard2.1</TargetFramework>
-        <Title>EcsRx.Plugins.Computeds</Title>
-        <Authors>Grofit (LP)</Authors>
-        <PackageLicenseUrl>https://github.com/ecsrx/ecsrx/blob/master/LICENSE</PackageLicenseUrl>
-        <PackageProjectUrl>https://github.com/ecsrx/ecsrx</PackageProjectUrl>
-        <Description>EcsRx plugin to add the notion of computeds to the system</Description>
-        <PackageTags>ecs rx reactive patterns ioc game-development xna monogame unity</PackageTags>
-        <LangVersion>8</LangVersion>
-    </PropertyGroup>
-
-    <ItemGroup>
-      <ProjectReference Include="..\EcsRx.Infrastructure\EcsRx.Infrastructure.csproj" />
-      <ProjectReference Include="..\EcsRx\EcsRx.csproj" />
-    </ItemGroup>
-
-    <ItemGroup>
-      <PackageReference Include="SystemsRx.Plugins.Computeds" Version="7.1.59" />
-    </ItemGroup>
-
-</Project>
