+++ src/EcsRx.Plugins.Transforms/EcsRx.Plugins.Transforms.csproj
@@ -13,8 +13,8 @@
     </PropertyGroup>
 
     <ItemGroup>
-      <PackageReference Include="SystemsRx.Infrastructure" Version="7.1.59" />
-      <PackageReference Include="SystemsRx.Plugins.Transforms" Version="7.1.59" />
+      <PackageReference Include="SystemsRx.Infrastructure" Version="8.0.62" />
+      <PackageReference Include="SystemsRx.Plugins.Transforms" Version="8.0.62" />
     </ItemGroup>
 
     <ItemGroup>
