+++ src/EcsRx.Tests/EcsRx.Tests.csproj
@@ -7,7 +7,7 @@
   <ItemGroup>
     <PackageReference Include="Microsoft.NET.Test.Sdk" Version="15.5.0" />
     <PackageReference Include="NSubstitute" Version="5.1.0" />
-    <PackageReference Include="SystemsRx.ReactiveData" Version="7.1.59" />
+    <PackageReference Include="SystemsRx.ReactiveData" Version="8.0.62" />
     <PackageReference Include="xunit" Version="2.3.1" />
     <PackageReference Include="xunit.runner.visualstudio" Version="2.3.1" />
     <DotNetCliToolReference Include="dotnet-xunit" Version="2.3.1" />
@@ -20,7 +20,6 @@
       <Name>EcsRx.Infrastructure</Name>
     </ProjectReference>
     <ProjectReference Include="..\EcsRx.Plugins.Batching\EcsRx.Plugins.Batching.csproj" />
-    <ProjectReference Include="..\EcsRx.Plugins.Computeds\EcsRx.Plugins.Computeds.csproj" />
     <ProjectReference Include="..\EcsRx.Plugins.GroupBinding\EcsRx.Plugins.GroupBinding.csproj" />
     <ProjectReference Include="..\EcsRx.Plugins.ReactiveSystems\EcsRx.Plugins.ReactiveSystems.csproj" />
     <ProjectReference Include="..\EcsRx.Plugins.Views\EcsRx.Plugins.Views.csproj" />
