+++ src/EcsRx.Tests/EcsRx/Computeds/ComputedCollectionFromGroupTests.cs
@@ -5,29 +5,34 @@ using System.Reactive.Subjects;
 using EcsRx.Entities;
 using EcsRx.Extensions;
 using EcsRx.Groups.Observable;
-using SystemsRx.MicroRx.Extensions;
+using EcsRx.Tests.EcsRx.Computeds.Models;
 using EcsRx.Tests.Models;
-using EcsRx.Tests.Plugins.Computeds.Models;
 using NSubstitute;
+using SystemsRx.MicroRx.Extensions;
 using Xunit;
 
-namespace EcsRx.Tests.Plugins.Computeds
+namespace EcsRx.Tests.EcsRx.Computeds
 {
-    public class ComputedFromGroupTests
-    {
+    public class ComputedCollectionFromGroupTests
+    {       
         [Fact]
         public void should_populate_on_creation()
         {
             var fakeEntity1 = Substitute.For<IEntity>();
+            fakeEntity1.Id.Returns(1);
             fakeEntity1.HasComponent<TestComponentThree>().Returns(false);
             
             var fakeEntity2 = Substitute.For<IEntity>();
+            fakeEntity2.Id.Returns(2);
             fakeEntity2.HasComponent<TestComponentThree>().Returns(true);
 
             var fakeEntity3 = Substitute.For<IEntity>();
+            fakeEntity2.Id.Returns(3);
             fakeEntity3.HasComponent<TestComponentThree>().Returns(true);
 
-            var expectedData = new List<IEntity> {fakeEntity2, fakeEntity3}.Average(x => x.GetHashCode());
+            var expectedData = new List<IEntity> {fakeEntity2, fakeEntity3}
+                .Select(x => x.GetHashCode())
+                .OrderBy(x => x);
 
             var fakeEntities = new List<IEntity> {fakeEntity1, fakeEntity2, fakeEntity3};
             
@@ -36,25 +41,30 @@ namespace EcsRx.Tests.Plugins.Computeds
             mockObservableGroup.OnEntityRemoving.Returns(Observable.Empty<IEntity>());
             mockObservableGroup.GetEnumerator().Returns(x => fakeEntities.GetEnumerator());
             
-            var computedGroupData = new TestComputedFromGroup(mockObservableGroup);
-            Assert.Equal(expectedData, computedGroupData.CachedData);            
+            var computedGroupData = new TestComputedCollectionFromGroup(mockObservableGroup);
+            Assert.Equal(expectedData, computedGroupData.Value);            
         }
         
         [Fact]
         public void should_refresh_when_entities_added_and_value_requested()
         {
             var fakeEntity1 = Substitute.For<IEntity>();
+            fakeEntity1.Id.Returns(1);
             fakeEntity1.HasComponent<TestComponentThree>().Returns(false);
             
             var fakeEntity2 = Substitute.For<IEntity>();
+            fakeEntity2.Id.Returns(2);
             fakeEntity2.HasComponent<TestComponentThree>().Returns(true);
 
             var fakeEntity3 = Substitute.For<IEntity>();
+            fakeEntity2.Id.Returns(3);
             fakeEntity3.HasComponent<TestComponentThree>().Returns(true);
 
-            var expectedData = fakeEntity3.GetHashCode();
+            var expectedData = new List<IEntity> {fakeEntity2, fakeEntity3}
+                .Select(x => x.GetHashCode())
+                .OrderBy(x => x);
 
-            var fakeEntities = new List<IEntity> {fakeEntity1, fakeEntity2, fakeEntity3};
+            var fakeEntities = new List<IEntity> {fakeEntity1, fakeEntity2};
             
             var mockObservableGroup = Substitute.For<IObservableGroup>();
 
@@ -64,9 +74,9 @@ namespace EcsRx.Tests.Plugins.Computeds
 
             mockObservableGroup.GetEnumerator().Returns(x => fakeEntities.GetEnumerator());
             
-            var computedGroupData = new TestComputedFromGroup(mockObservableGroup);
+            var computedGroupData = new TestComputedCollectionFromGroup(mockObservableGroup);
 
-            fakeEntities.Remove(fakeEntity2);
+            fakeEntities.Add(fakeEntity3);
             addedEvent.OnNext(null);
             
             var actualData = computedGroupData.Value;
@@ -78,81 +88,96 @@ namespace EcsRx.Tests.Plugins.Computeds
         public void should_refresh_when_entities_removed_and_value_requested()
         {
             var fakeEntity1 = Substitute.For<IEntity>();
+            fakeEntity1.Id.Returns(1);
             fakeEntity1.HasComponent<TestComponentThree>().Returns(false);
             
             var fakeEntity2 = Substitute.For<IEntity>();
+            fakeEntity2.Id.Returns(2);
             fakeEntity2.HasComponent<TestComponentThree>().Returns(true);
 
             var fakeEntity3 = Substitute.For<IEntity>();
+            fakeEntity2.Id.Returns(3);
             fakeEntity3.HasComponent<TestComponentThree>().Returns(true);
 
-            var expectedData = fakeEntity3.GetHashCode();
+            var expectedData = new[] {fakeEntity2.GetHashCode()};
 
-            var fakeEntities = new List<IEntity> {fakeEntity1, fakeEntity2, fakeEntity3};
+            var fakeEntities = new List<IEntity> {fakeEntity1, fakeEntity2};
             
             var mockObservableGroup = Substitute.For<IObservableGroup>();
 
-            var removedEvent = new Subject<IEntity>();
+            var removingEntity = new Subject<IEntity>();
             mockObservableGroup.OnEntityAdded.Returns(Observable.Empty<IEntity>());
-            mockObservableGroup.OnEntityRemoving.Returns(removedEvent);
+            mockObservableGroup.OnEntityRemoving.Returns(removingEntity);
 
             mockObservableGroup.GetEnumerator().Returns(x => fakeEntities.GetEnumerator());
             
-            var computedGroupData = new TestComputedFromGroup(mockObservableGroup);
+            var computedGroupData = new TestComputedCollectionFromGroup(mockObservableGroup);
 
-            fakeEntities.Remove(fakeEntity2);
-            removedEvent.OnNext(null);
+            fakeEntities.Remove(fakeEntity3);
+            removingEntity.OnNext(fakeEntity3);
             
             var actualData = computedGroupData.Value;
 
-            Assert.Equal(expectedData, actualData);         
+            Assert.Equal(expectedData, actualData);               
         }
         
         [Fact]
         public void should_refresh_on_trigger_and_value_requested()
         {
             var fakeEntity1 = Substitute.For<IEntity>();
+            fakeEntity1.Id.Returns(1);
             fakeEntity1.HasComponent<TestComponentThree>().Returns(false);
             
             var fakeEntity2 = Substitute.For<IEntity>();
+            fakeEntity2.Id.Returns(2);
             fakeEntity2.HasComponent<TestComponentThree>().Returns(true);
 
             var fakeEntity3 = Substitute.For<IEntity>();
+            fakeEntity2.Id.Returns(3);
             fakeEntity3.HasComponent<TestComponentThree>().Returns(true);
 
-            var expectedData = fakeEntity3.GetHashCode();
+            var expectedData = new List<IEntity> {fakeEntity2, fakeEntity3}
+                .Select(x => x.GetHashCode())
+                .OrderBy(x => x);
 
-            var fakeEntities = new List<IEntity> {fakeEntity1, fakeEntity2, fakeEntity3};
+            var fakeEntities = new List<IEntity> {fakeEntity1, fakeEntity2};
             
             var mockObservableGroup = Substitute.For<IObservableGroup>();
-            mockObservableGroup.OnEntityAdded.Returns(Observable.Empty<IEntity>());
+
+            var addedEvent = new Subject<IEntity>();
+            mockObservableGroup.OnEntityAdded.Returns(addedEvent);
             mockObservableGroup.OnEntityRemoving.Returns(Observable.Empty<IEntity>());
 
             mockObservableGroup.GetEnumerator().Returns(x => fakeEntities.GetEnumerator());
             
-            var computedGroupData = new TestComputedFromGroup(mockObservableGroup);
+            var computedGroupData = new TestComputedCollectionFromGroup(mockObservableGroup);
 
-            fakeEntities.Remove(fakeEntity2);
+            fakeEntities.Add(fakeEntity3);
             computedGroupData.ManuallyRefresh.OnNext(true);
-                       
+            
             var actualData = computedGroupData.Value;
 
-            Assert.Equal(expectedData, actualData);         
+            Assert.Equal(expectedData, actualData);              
         }
         
         [Fact]
-        public void should_not_refresh_value_with_no_subs_and_value_requested()
+        public void should_not_refresh_value_with_no_subs_events_or_triggers_and_value_requested()
         {
             var fakeEntity1 = Substitute.For<IEntity>();
+            fakeEntity1.Id.Returns(1);
             fakeEntity1.HasComponent<TestComponentThree>().Returns(false);
             
             var fakeEntity2 = Substitute.For<IEntity>();
+            fakeEntity2.Id.Returns(2);
             fakeEntity2.HasComponent<TestComponentThree>().Returns(true);
 
             var fakeEntity3 = Substitute.For<IEntity>();
+            fakeEntity2.Id.Returns(3);
             fakeEntity3.HasComponent<TestComponentThree>().Returns(true);
 
-            var expectedData = new [] { fakeEntity2.GetHashCode(), fakeEntity3.GetHashCode() }.Average();
+            var expectedData = new List<IEntity> {fakeEntity2, fakeEntity3}
+                .Select(x => x.GetHashCode())
+                .OrderBy(x => x);
 
             var fakeEntities = new List<IEntity> {fakeEntity1, fakeEntity2, fakeEntity3};
             
@@ -162,7 +187,7 @@ namespace EcsRx.Tests.Plugins.Computeds
 
             mockObservableGroup.GetEnumerator().Returns(x => fakeEntities.GetEnumerator());
             
-            var computedGroupData = new TestComputedFromGroup(mockObservableGroup);
+            var computedGroupData = new TestComputedCollectionFromGroup(mockObservableGroup);
 
             fakeEntities.Remove(fakeEntity2);            
             var actualData = computedGroupData.Value;
@@ -171,15 +196,18 @@ namespace EcsRx.Tests.Plugins.Computeds
         }
         
         [Fact]
-        public void should_not_refresh_cached_data_on_change_notified_but_no_subs_without_value_request()
+        public void should_not_refresh_cached_data_when_change_notified_but_no_value_request()
         {
             var fakeEntity1 = Substitute.For<IEntity>();
+            fakeEntity1.Id.Returns(1);
             fakeEntity1.HasComponent<TestComponentThree>().Returns(false);
             
             var fakeEntity2 = Substitute.For<IEntity>();
+            fakeEntity2.Id.Returns(2);
             fakeEntity2.HasComponent<TestComponentThree>().Returns(true);
 
             var fakeEntity3 = Substitute.For<IEntity>();
+            fakeEntity2.Id.Returns(3);
             fakeEntity3.HasComponent<TestComponentThree>().Returns(true);
 
             var fakeEntities = new List<IEntity> {fakeEntity1, fakeEntity2, fakeEntity3};
@@ -197,26 +225,30 @@ namespace EcsRx.Tests.Plugins.Computeds
             var computedGroupData = new TestComputedFromGroup(mockObservableGroup);
             var expectedOutput = computedGroupData.CachedData;
 
-            addingSubject.OnNext(null);           
-            removingSubject.OnNext(null);
+            addingSubject.OnNext(null);
             computedGroupData.ManuallyRefresh.OnNext(true);
 
             Assert.Equal(expectedOutput, computedGroupData.CachedData);         
         }
         
         [Fact]
-        public void should_refresh_cached_data_on_change_notified_with_active_subs_without_value_request()
+        public void should_refresh_cached_data_when_change_notified_with_subs_but_no_value_requested()
         {
             var fakeEntity1 = Substitute.For<IEntity>();
+            fakeEntity1.Id.Returns(1);
             fakeEntity1.HasComponent<TestComponentThree>().Returns(false);
             
             var fakeEntity2 = Substitute.For<IEntity>();
+            fakeEntity2.Id.Returns(2);
             fakeEntity2.HasComponent<TestComponentThree>().Returns(true);
 
             var fakeEntity3 = Substitute.For<IEntity>();
+            fakeEntity2.Id.Returns(3);
             fakeEntity3.HasComponent<TestComponentThree>().Returns(true);
 
-            var expectedData = new [] { fakeEntity2.GetHashCode(), fakeEntity3.GetHashCode() }.Average();
+            var expectedData = new List<IEntity> {fakeEntity2, fakeEntity3}
+                .Select(x => x.GetHashCode())
+                .OrderBy(x => x);
 
             var fakeEntities = new List<IEntity> {fakeEntity1, fakeEntity2, fakeEntity3};
             
@@ -230,21 +262,20 @@ namespace EcsRx.Tests.Plugins.Computeds
 
             mockObservableGroup.GetEnumerator().Returns(x => fakeEntities.GetEnumerator());
             
-            var computedGroupData = new TestComputedFromGroup(mockObservableGroup);
+            var computedGroupData = new TestComputedCollectionFromGroup(mockObservableGroup);
             computedGroupData.Subscribe(x => {});
 
             fakeEntities.Remove(fakeEntity2);
-            removingSubject.OnNext(null);
-            Assert.NotEqual(expectedData, computedGroupData.CachedData);
+            removingSubject.OnNext(fakeEntity2);
+            Assert.False(expectedData.SequenceEqual(computedGroupData.Value));
             
             fakeEntities.Add(fakeEntity2);
-            addingSubject.OnNext(null);  
-            Assert.Equal(expectedData, computedGroupData.CachedData);
+            addingSubject.OnNext(fakeEntity2);  
+            Assert.True(expectedData.SequenceEqual(computedGroupData.Value));
 
             fakeEntities.Remove(fakeEntity2);
             computedGroupData.ManuallyRefresh.OnNext(true);
-
-            Assert.NotEqual(expectedData, computedGroupData.CachedData);         
+            Assert.False(expectedData.SequenceEqual(computedGroupData.Value));      
         }
     }
 }
