+++ src/EcsRx.Tests/EcsRx/Computeds/Models/TestComputedFromGroup.cs
@@ -1,12 +1,12 @@
 using System;
 using System.Linq;
+using EcsRx.Computeds;
 using EcsRx.Extensions;
 using EcsRx.Groups.Observable;
-using SystemsRx.MicroRx.Subjects;
-using EcsRx.Plugins.Computeds;
 using EcsRx.Tests.Models;
+using SystemsRx.MicroRx.Subjects;
 
-namespace EcsRx.Tests.Plugins.Computeds.Models
+namespace EcsRx.Tests.EcsRx.Computeds.Models
 {
     public class TestComputedFromGroup : ComputedFromGroup<double>
     {
