+++ src/EcsRx.Tests/EcsRx/Computeds/Models/TestComputedGroup.cs
@@ -1,12 +1,12 @@
 ﻿using System;
+using EcsRx.Computeds.Groups;
 using EcsRx.Entities;
 using EcsRx.Extensions;
 using EcsRx.Groups.Observable;
-using SystemsRx.MicroRx.Subjects;
-using EcsRx.Plugins.Computeds.Groups;
 using EcsRx.Tests.Models;
+using SystemsRx.MicroRx.Subjects;
 
-namespace EcsRx.Tests.Plugins.Computeds.Models
+namespace EcsRx.Tests.EcsRx.Computeds.Models
 {
     public class TestComputedGroup : ComputedGroup
     {
