+++ src/EcsRx.sln
@@ -25,8 +25,6 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "EcsRx.Plugins.ReactiveSyste
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "EcsRx.Plugins.Views", "EcsRx.Plugins.Views\EcsRx.Plugins.Views.csproj", "{F7B3C0BA-F801-4667-A46A-CDB8887BE08B}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "EcsRx.Plugins.Computeds", "EcsRx.Plugins.Computeds\EcsRx.Plugins.Computeds.csproj", "{5CB6728F-7537-4C73-9557-52B4C6F9791A}"
-EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "EcsRx.Plugins.Persistence", "EcsRx.Plugins.Persistence\EcsRx.Plugins.Persistence.csproj", "{D63A4C86-6515-4A3C-9473-45947D34DE21}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "EcsRx.Plugins.GroupBinding", "EcsRx.Plugins.GroupBinding\EcsRx.Plugins.GroupBinding.csproj", "{3DD20735-C1F5-4033-A909-2E1BD9EA075F}"
@@ -69,10 +67,6 @@ Global
 		{F7B3C0BA-F801-4667-A46A-CDB8887BE08B}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{F7B3C0BA-F801-4667-A46A-CDB8887BE08B}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{F7B3C0BA-F801-4667-A46A-CDB8887BE08B}.Release|Any CPU.Build.0 = Release|Any CPU
-		{5CB6728F-7537-4C73-9557-52B4C6F9791A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{5CB6728F-7537-4C73-9557-52B4C6F9791A}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{5CB6728F-7537-4C73-9557-52B4C6F9791A}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{5CB6728F-7537-4C73-9557-52B4C6F9791A}.Release|Any CPU.Build.0 = Release|Any CPU
 		{D63A4C86-6515-4A3C-9473-45947D34DE21}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{D63A4C86-6515-4A3C-9473-45947D34DE21}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{D63A4C86-6515-4A3C-9473-45947D34DE21}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -101,7 +95,6 @@ Global
 		{D238EA0D-7A84-42FB-9E74-B56E463011F0} = {48050703-1E4C-4179-9539-90212854B3F7}
 		{92E2383A-2386-47DF-A0D7-83348ADBFBDB} = {48050703-1E4C-4179-9539-90212854B3F7}
 		{F7B3C0BA-F801-4667-A46A-CDB8887BE08B} = {48050703-1E4C-4179-9539-90212854B3F7}
-		{5CB6728F-7537-4C73-9557-52B4C6F9791A} = {48050703-1E4C-4179-9539-90212854B3F7}
 		{D63A4C86-6515-4A3C-9473-45947D34DE21} = {48050703-1E4C-4179-9539-90212854B3F7}
 		{3DD20735-C1F5-4033-A909-2E1BD9EA075F} = {48050703-1E4C-4179-9539-90212854B3F7}
 		{4037E761-E07D-42B1-91B8-69798CF0B678} = {4197CF20-C162-449D-A3A5-D50BF9476796}
