+++ src/EcsRx/Computeds/Groups/IComputedGroup.cs
@@ -1,6 +1,6 @@
 ﻿using EcsRx.Groups.Observable;
 
-namespace EcsRx.Plugins.Computeds.Groups
+namespace EcsRx.Computeds.Groups
 {
     public interface IComputedGroup : IObservableGroup
     {
